# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Cricket Box API
  version: 1.0.0
  description: |
    ## Cricket Box - Digital Box Cricket Management Platform
    
    Complete API documentation for Cricket Box backend system.
    
    ### Features
    - User Authentication (Email/Password & OTP)
    - Room Management
    - Team Setup
    - Live Match Scoring
    - Statistics & Leaderboards
    - Admin Dashboard
    
  contact:
    name: Cricket Box Support
    email: support@cricketbox.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.cricketbox.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errorCode:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
        meta:
          type: object
          properties:
            pagination:
              type: object
              properties:
                currentPage:
                  type: integer
                itemsPerPage:
                  type: integer
                totalItems:
                  type: integer
                totalPages:
                  type: integer
                hasNextPage:
                  type: boolean
                hasPrevPage:
                  type: boolean
  responses:
...
    ForbiddenError:
      description: User does not have permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: Get all users (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, host, player, umpire, viewer]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, blocked, pending_verification]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /users/leaderboard:
    get:
      tags: [Users]
      summary: Get leaderboard
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [runs, wickets, matches, wins, sixes, fours, catches]
            default: runs
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Leaderboard data
  /users/{userId}:
    get:
      tags: [Users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags: [Users]
      summary: Update user profile
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                phone:
                  type: string
                avatar:
                  type: string
                  format: uri
      responses:
        '200':
          description: Profile updated
        '403':
          description: Can only update own profile
  /users/{userId}/statistics:
    get:
      tags: [Users]
      summary: Get user statistics
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User statistics

  # Room Endpoints
...
  /admin/users:
    get:
      tags: [Admin]
      summary: Get all users with admin view
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of users
...